sort validation function into separate files
