+++ helper/validation/strings_test.go
@@ -1,6 +1,7 @@
 package validation
 
 import (
+	`regexp`
 	"testing"
 )
 
@@ -273,4 +274,193 @@ func TestValidationStringIsBase64(t *testing.T) {
 			}
 		})
 	}
+}
+
+func TestValidationStringInSlice(t *testing.T) {
+	runTestCases(t, []testCase{
+		{
+			val: "ValidValue",
+			f:   StringInSlice([]string{"ValidValue", "AnotherValidValue"}, false),
+		},
+		// ignore case
+		{
+			val: "VALIDVALUE",
+			f:   StringInSlice([]string{"ValidValue", "AnotherValidValue"}, true),
+		},
+		{
+			val:         "VALIDVALUE",
+			f:           StringInSlice([]string{"ValidValue", "AnotherValidValue"}, false),
+			expectedErr: regexp.MustCompile("expected [\\w]+ to be one of \\[ValidValue AnotherValidValue\\], got VALIDVALUE"),
+		},
+		{
+			val:         "InvalidValue",
+			f:           StringInSlice([]string{"ValidValue", "AnotherValidValue"}, false),
+			expectedErr: regexp.MustCompile("expected [\\w]+ to be one of \\[ValidValue AnotherValidValue\\], got InvalidValue"),
+		},
+		{
+			val:         1,
+			f:           StringInSlice([]string{"ValidValue", "AnotherValidValue"}, false),
+			expectedErr: regexp.MustCompile("expected type of [\\w]+ to be string"),
+		},
+	})
+}
+
+func TestValidationStringMatch(t *testing.T) {
+	runTestCases(t, []testCase{
+		{
+			val: "foobar",
+			f:   StringMatch(regexp.MustCompile(".*foo.*"), ""),
+		},
+		{
+			val:         "bar",
+			f:           StringMatch(regexp.MustCompile(".*foo.*"), ""),
+			expectedErr: regexp.MustCompile("expected value of [\\w]+ to match regular expression " + regexp.QuoteMeta(`".*foo.*"`)),
+		},
+		{
+			val:         "bar",
+			f:           StringMatch(regexp.MustCompile(".*foo.*"), "value must contain foo"),
+			expectedErr: regexp.MustCompile("invalid value for [\\w]+ \\(value must contain foo\\)"),
+		},
+	})
+}
+
+func TestValidationStringDoesNotMatch(t *testing.T) {
+	runTestCases(t, []testCase{
+		{
+			val: "foobar",
+			f:   StringDoesNotMatch(regexp.MustCompile(".*baz.*"), ""),
+		},
+		{
+			val:         "bar",
+			f:           StringDoesNotMatch(regexp.MustCompile(".*bar.*"), ""),
+			expectedErr: regexp.MustCompile("expected value of [\\w]+ to not match regular expression " + regexp.QuoteMeta(`".*bar.*"`)),
+		},
+		{
+			val:         "bar",
+			f:           StringDoesNotMatch(regexp.MustCompile(".*bar.*"), "value must not contain foo"),
+			expectedErr: regexp.MustCompile("invalid value for [\\w]+ \\(value must not contain foo\\)"),
+		},
+	})
+}
+
+
+func TestValidateJsonString(t *testing.T) {
+	type testCases struct {
+		Value    string
+		ErrCount int
+	}
+
+	invalidCases := []testCases{
+		{
+			Value:    `{0:"1"}`,
+			ErrCount: 1,
+		},
+		{
+			Value:    `{'abc':1}`,
+			ErrCount: 1,
+		},
+		{
+			Value:    `{"def":}`,
+			ErrCount: 1,
+		},
+		{
+			Value:    `{"xyz":[}}`,
+			ErrCount: 1,
+		},
+	}
+
+	for _, tc := range invalidCases {
+		_, errors := ValidateJsonString(tc.Value, "json")
+		if len(errors) != tc.ErrCount {
+			t.Fatalf("Expected %q to trigger a validation error.", tc.Value)
+		}
+	}
+
+	validCases := []testCases{
+		{
+			Value:    ``,
+			ErrCount: 0,
+		},
+		{
+			Value:    `{}`,
+			ErrCount: 0,
+		},
+		{
+			Value:    `{"abc":["1","2"]}`,
+			ErrCount: 0,
+		},
+	}
+
+	for _, tc := range validCases {
+		_, errors := ValidateJsonString(tc.Value, "json")
+		if len(errors) != tc.ErrCount {
+			t.Fatalf("Expected %q not to trigger a validation error.", tc.Value)
+		}
+	}
+}
+
+func TestValidateListUniqueStrings(t *testing.T) {
+	runTestCases(t, []testCase{
+		{
+			val: []interface{}{"foo", "bar"},
+			f:   ValidateListUniqueStrings,
+		},
+		{
+			val:         []interface{}{"foo", "bar", "foo"},
+			f:           ValidateListUniqueStrings,
+			expectedErr: regexp.MustCompile("duplicate entry - foo"),
+		},
+		{
+			val:         []interface{}{"foo", "bar", "foo", "baz", "bar"},
+			f:           ValidateListUniqueStrings,
+			expectedErr: regexp.MustCompile("duplicate entry - (?:foo|bar)"),
+		},
+	})
+}
+
+
+func TestStringDoesNotContainAny(t *testing.T) {
+	chars := "|:/"
+
+	validStrings := []string{
+		"HelloWorld",
+		"ABC_*&^%123",
+	}
+	for _, v := range validStrings {
+		_, errors := StringDoesNotContainAny(chars)(v, "name")
+		if len(errors) != 0 {
+			t.Fatalf("%q should not contain any of %q", v, chars)
+		}
+	}
+
+	invalidStrings := []string{
+		"Hello/World",
+		"ABC|123",
+		"This will fail:",
+		chars,
+	}
+	for _, v := range invalidStrings {
+		_, errors := StringDoesNotContainAny(chars)(v, "name")
+		if len(errors) == 0 {
+			t.Fatalf("%q should contain one of %q", v, chars)
+		}
+	}
+}
+
+
+
+
+
+func TestValidationRegexp(t *testing.T) {
+	runTestCases(t, []testCase{
+		{
+			val: ".*foo.*",
+			f:   ValidateRegexp,
+		},
+		{
+			val:         "foo(bar",
+			f:           ValidateRegexp,
+			expectedErr: regexp.MustCompile(regexp.QuoteMeta("error parsing regexp: missing closing ): `foo(bar`")),
+		},
+	})
 }
